refactor: 💡 add index.ts
