+++ src/config/index.ts
@@ -1,0 +1,4 @@
+import configuration from './configuration';
+import IConfig from './IConfig';
+
+export { configuration, IConfig };
