+++ src/libs/errorHandler.ts
@@ -1,5 +1,5 @@
+import { INotFoundError } from './routes';
 import IErrorHandlerResponse from './IErrorHandlerResponse';
-import INotFoundError from './routes/INotFoundError';
 import { Request, Response, NextFunction } from 'express';
 
 const ErrorHandler = (err: INotFoundError, req: Request, res: Response, next: NextFunction) => {
